Proposal: change news order
Change the news list to have the most recent at the top, so visitors to the site see it first. Reasons:
(1) First time I visited the page, I momentarily assumed nothing had happened on the project since 2017. Presume others may make the same error.
(2) For new or returning visitors, news from a year or two ago isn't as relevant as the most recent announcements, so offer the newest first.