Call out Browsersync `<body>` requirement on Getting Started page
This page currently says, "Make a change to your template files and save them—Eleventy using BrowserSync will refresh the browser with your new changes automatically." This does not, however, apply to the README.md file just created in the previous step on this page, since it doesn't include a <body> tag. Explain that caveat here, not just on the subsequent page of the docs.