add a link to markdown shortcode plugin
Hey! Not sure what the process is to end up on this documentation page. I've got the plugin tagged on NPM with eleventy-plugin, so if it's not appropriate to add directly to the docs, not a problem. Just figured I'd take a stab at adding it since some people seem to be grabbing it.