Create Lowtus.json
Example of websites made with Eleventy. This a migration from Jekyll to Eleventy with a try of keeping compatibility between each other (LiquidJS). The code is not open because the architecture is not stable and not usable.
Lowtus is a non profit action for helping people to see websites energy impact.