Suggestion: Live refresh 404 pages on change
When using `eleventy --serve"` and the current documentation, you have to cancel the session and start again when you make changes to the 404 page. 

I would like to suggest the documentation to be updated to move `fs.readFileSync('_site/404.html');` into the `addMiddleware` function. This allows the 404 page to live refresh as changes are made to a markdown file. The downsides of this are that on every time browser-sync is ran on a 404 route it will read the `'_site/404.html'` file and update what is sent back to the user.

Let me know what you think? It sounds like an experience versus performance trade off.