Update debugging.md
I struggled for a while to get debugging running on windows. I found this issue for another project that helped me https://github.com/getgauge/taiko/issues/1065. The formatting of the command to set the debug value didn't work for me, but using the formatting from that issue solved my problem. So consider adding this to the document to save anyone else from having to dig around for the same information.