Update collections.md
Added a warning about 'tags' terminology. As quoted here by Josh Cunningham : https://www.joshcanhelp.com/taking-wordpress-to-eleventy/
it may be confusing coming from Wordpress or any blogging platform : 
Things can get a bit confusing here as we're going to start to overlap terminology a bit between the two platforms. In Eleventy, tags are used to construct collections of content. In WordPress, tags are just a type of taxonomy with no hierarchy.

As I'm not an native english speaker, I'm not sure about the way of writing this warning...