Add the note about a multiple word tag
From my point of view, it is worth adding the note about using a
multiple word tag.

It would be extremely useful for newcomers who play with the official
starter project. That project contains a lot of multiple word tags and
looking at the current doc page it's not clear how to manage a post with
a tag like `number 2` or `second tag`.