Create ribomation.json
Many thanks for the 11ty SSG. Just converted our site from Vue/Nuxt to 11ty. Now it's blazingly fast compared to before and its rising in SEO ranking as well. The hard lesson learnt is that a SPA site with SSR or SSG, simply do not fly when it comes to SEO and site speed. 

Before Vue/Nuxt, we used Angular with SSR, which was much worser than the Nuxt version. Before Angular, we used a terrible hack based on Gulp v3, with features not far from 11ty, but of course not with all bells & whistles and generality of 11ty. So, in a remote sense, we are back to where we were three years ago, but on a higher excited energy level.