Cool URIs do not change!
In this section about permalinks, it states "What to leave out…
File name extension. … The canonical way of making links to the W3C site doesn't use the extension."
Immediately before this quote, it links to `https://www.w3.org/Provider/Style/URI.html`.

Ironic? (it's fine, the URI was probably copy-pasted, but still an error!)