Update rss.md - add specifics for feed template name and location
Add specific instructions for the sample feed template filename and location.

The existing instructions are unclear as to exactly where to save the feed template file to, and what it should be named. 

It took two hours Googling and trying different options until figuring out that it's not "Place the file anywhere in your project" but "anywhere in the src directory". The existing docs don't say what this file should be named which could be confusing.