Mention that the web component only works in webc templates
This was confusing to me; I didn't even realise that there was a "webc" format for templates. While the docs do technically say this ("Using Eleventy Image in WebC"), that's only really clear to someone who knows already what it means. A gentle pointer to say "if you're trying to use this in some other template format, you need to use the shortcodes" will help people who aren't aware of this to understand why it's not working for them. The wording here might not be right, of course; this is a proposal to make the docs a little more clear and helpful, and if that's better done by wording it slightly differently or more precisely, that's fine.