Update index.webc
At the 4th step of the document, the  out-file command in windows is Power Shell specific. But Power Shell has a default execution policy to restrict the running of JavaScript, which would cause an error when running Eleventy in the next step.
If a solution can be provided here(use a command to set the execution policy), the whole experience of trying running Eleventy would be smoother and more beginner-friendly.